27-4-18 9:13

+++ src/com/controller/AlertController.java
@@ -1,5 +1,6 @@
 package com.controller;
 
+import java.security.Principal;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
@@ -23,9 +24,11 @@ import org.springframework.web.bind.annotation.RestController;
 import com.model.Alert;
 import com.model.ControlReception;
 import com.model.Phase;
+import com.model.Utilisateur;
 import com.service.AlertService;
 import com.service.ControlReceptionService;
 import com.service.PhaseService;
+import com.service.UtilisateurService;
 import com.wrapper.AlertWrapper;
 
 @Controller
@@ -34,6 +37,14 @@ public class AlertController {
 	private AlertService alertService;
 	private ControlReceptionService controlReceptionService;
 	private PhaseService phaseService;
+	private UtilisateurService utilisateurService;
+	
+	
+	@Autowired
+	public void setUtilisateurService(UtilisateurService us){
+		utilisateurService = us;
+	}
+	
 
 	@Autowired
 	@Qualifier(value="phaseService")
@@ -55,17 +66,28 @@ public class AlertController {
 	
 	@RequestMapping(value="alerts/store",method=RequestMethod.POST)
 	public String store(@RequestParam("type")String type,
-			@RequestParam("id")Long id) {
+			@RequestParam("id")Long id,Principal p) {
+		
 		ControlReception controlReception = controlReceptionService.find(id);
 		if(controlReception != null)
 		{
+			 List<String> access = new ArrayList<String>();
+			    access.add("Technicient qualite");
+			    access.add("Responsable qualite");
+			    access.add("Injenieur qualite");
+			    access.add("Coordinatrice");
+			    Utilisateur u = utilisateurService.getUtilisateurByUsername(p.getName());
+			    
+			    if(!access.contains(u.getUtilisateurRoles().get(0).getRole())) {
+			    	return "denied";
+			    }
+			    
 			Alert alert = new Alert();
 			alert.setControlReception(controlReception);
 			alert.setDate_creation(new Date());
 			alert.setStatus("open");
 			alert.setType(type);
 			alertService.save(alert);
-			
 		}
 	return  "redirect:/controlR/index";
 		
@@ -73,22 +95,44 @@ public class AlertController {
 	
 	@RequestMapping(value="alerts/{num}/show",method=RequestMethod.GET)
 
-	public String show(@PathVariable("num")Long num,ModelMap map)
+	public String show(@PathVariable("num")Long num,ModelMap map,Principal p)
 	{
+		List<String> access = new ArrayList<String>();
+	    access.add("Technicient qualite");
+	    access.add("Responsable qualite");
+	    access.add("Injenieur qualite");
+	    access.add("Coordinatrice");
+	    Utilisateur u = utilisateurService.getUtilisateurByUsername(p.getName());
+	    
+	    if(!access.contains(u.getUtilisateurRoles().get(0).getRole())) {
+	    	return "denied";
+	    }
 		Alert alert = alertService.find(num);
 		List<Phase> phases = phaseService.fetchAll();
 		map.addAttribute("alert",alert);
 		map.addAttribute("phases",phases);
+		map.addAttribute("utilisateur",u);
 		return "alert/show";
 		
 
 	}
 
 	@RequestMapping(value="alerts/{num}/edit",method=RequestMethod.GET)
-	public String edit(@PathVariable("num")Long num,ModelMap map)
+	public String edit(@PathVariable("num")Long num,ModelMap map,Principal p)
 	{
+		List<String> access = new ArrayList<String>();
+	    access.add("Technicient qualite");
+	    access.add("Responsable qualite");
+	    access.add("Injenieur qualite");
+	    access.add("Coordinatrice");
+	    Utilisateur u = utilisateurService.getUtilisateurByUsername(p.getName());
+	    
+	    if(!access.contains(u.getUtilisateurRoles().get(0).getRole())) {
+	    	return "denied";
+	    }
 		Alert alert = alertService.find(num);
 		map.addAttribute("alert",alert);
+		map.addAttribute("utilisateur",u);
 		return "alert/edit";
 	}
 	
@@ -97,8 +141,18 @@ public class AlertController {
 			@RequestParam("phase")String phase,
 			@RequestParam("chargeback")Double chargeback,
 			@RequestParam("type")String type,
-			HttpServletRequest request)
+			HttpServletRequest request,Principal p)
 	{
+		List<String> access = new ArrayList<String>();
+	    access.add("Technicient qualite");
+	    access.add("Responsable qualite");
+	    access.add("Injenieur qualite");
+	    access.add("Coordinatrice");
+	    Utilisateur u = utilisateurService.getUtilisateurByUsername(p.getName());
+	    
+	    if(!access.contains(u.getUtilisateurRoles().get(0).getRole())) {
+	    	return "denied";
+	    }
 		Alert alert = alertService.find(numero);
 		alert.setChargeback(chargeback);
 		alert.setPhase(phase);
@@ -110,8 +164,18 @@ public class AlertController {
 	
 	
 	@RequestMapping(value="alerts/{num}/close",method=RequestMethod.GET)
-	public String close(@PathVariable("num")Long numero,HttpServletRequest request)
+	public String close(@PathVariable("num")Long numero,HttpServletRequest request,Principal p)
 	{
+		List<String> access = new ArrayList<String>();
+	    access.add("Technicient qualite");
+	    access.add("Responsable qualite");
+	    access.add("Injenieur qualite");
+	    access.add("Coordinatrice");
+	    Utilisateur u = utilisateurService.getUtilisateurByUsername(p.getName());
+	    
+	    if(!access.contains(u.getUtilisateurRoles().get(0).getRole())) {
+	    	return "denied";
+	    }
 		Alert alert = alertService.find(numero);
 		alert.setStatus("close");
 		alert.setDate_close(new Date());
@@ -120,8 +184,18 @@ public class AlertController {
 	}
 	
 	@RequestMapping(value="alerts/index",method=RequestMethod.GET)
-	public String index(ModelMap map)
+	public String index(ModelMap map,Principal p)
 	{
+		List<String> access = new ArrayList<String>();
+	    access.add("Technicient qualite");
+	    access.add("Responsable qualite");
+	    access.add("Injenieur qualite");
+	    access.add("Coordinatrice");
+	    Utilisateur u = utilisateurService.getUtilisateurByUsername(p.getName());
+	    
+	    if(!access.contains(u.getUtilisateurRoles().get(0).getRole())) {
+	    	return "denied";
+	    }
 		List<Alert> dangers = new ArrayList<>();
 		
 		List<Alert> alertsOpen = alertService.fetchAllOpen();
@@ -152,12 +226,24 @@ public class AlertController {
 		}
 		map.addAttribute("dangers",dangers);
 		map.addAttribute("items",items);
+		map.addAttribute("utilisateur",u);
 		return "alert/lister";
 	}
 	
 	@RequestMapping(value="alerts/{num}/delete",method=RequestMethod.GET)
-	public String destroy(@PathVariable("num")Long num)
+	public String destroy(@PathVariable("num")Long num,Principal p)
 	{
+
+		List<String> access = new ArrayList<String>();
+	    access.add("Technicient qualite");
+	    access.add("Responsable qualite");
+	    access.add("Injenieur qualite");
+	    access.add("Coordinatrice");
+	    Utilisateur u = utilisateurService.getUtilisateurByUsername(p.getName());
+	    
+	    if(!access.contains(u.getUtilisateurRoles().get(0).getRole())) {
+	    	return "denied";
+	    }
 		alertService.delete(num);
 		return "redirect:/alerts/index";
 	}
